Added README for meetings directory
Added README for the `meetings` directory where the yaml files for each
meeting is located. The README includes a link to the OpenStack Meeting
page on the wiki, a link to the meeting yaml files on OpenStack git
site, as well as brief explanation on the content inside a yaml file.

The description of the yaml content now uses | to perserve the "list"
structure. This section is further edited so that `*` are used to
format these bulleted list of agenda items instead of `-`.

Change-Id: Iab1af314dc979394bd4a61caccf902cc92bf3735