Added command line tool information to README.rst
Added explanations and example scenarios about the command line usage for
jobs.py to the README.rst file.  A few syntax errors found in the
README.rst was fixed as well. Visit the following link for a preview of
the README.rst:

  http://rst.ninjs.org/?n=de7d3bc022f4ba2417bd3de97e33c91c&theme=nature

Additionally, the YAML file extention has been changed to '.yaml.' The
name of the meetings have also been capitalized. Spaces were added
in front of agenda list items.

Lastly, a minor issue was fixed where the iCals files was missing the .
in the file extension. It now can handle both '.yaml' and '.yml'
extentions when writting '.ical' files.

Change-Id: Iff6805a690339303ec58bf19883dcfbbf3d37d72