Refactor schedule class into its own module
Previously, the schedule class was in the meeting module with the
meeting class. This change pulls schedule into its own module. This will
be easier for unit testing and will make it is easier for us to change
in the future if we want to incorporate the schedule class purpose into
the meeting object.

Change-Id: I3f03c2167997b39fadd362bd66a35cae570503ef