Refactor next_weekday into Meeting class
Previously, the next_weekday method was the only module level function
in the meeting module, and that was the only place it was used. This
change incorporates it into the Meeting class since it makes sense to
have it there since that's the only thing that uses it.

This also touches up the doc strings of next_weekday.

Change-Id: I3af654b480c1585ae19841a0baa39079c0ad0d83