Refactor Meeting init and enable unit tests
Previously, there were two load_meetings() methods. One was in util.py
and one was a member of the Meeting object. This commit combines the two
and makes it general enough for both cases.

Update tox.ini as well as .testr.conf so that we can run some basic unit
tests. An initial test has been added so there is something to run.

Change-Id: I44fd693f1a4c0a0efcf99d935123108f22afea36