Fix default paths for meetings and ical dir
Previously, the iCal directory and the meetings directory defaulted
to relative paths. This change defaults those paths to a full path name
for both.

This change also adds the icals directory to the .gitignore of the
project. This way when users generate icals and use the default
settings, they aren't adding icals to the project repository.

Change-Id: If5d62d2c74a7edebc00d58cf210e47c45ec2ea7b