Rename 'arbiter' module to 'yaml2ical'
We'll need to upload this module to PyPI in order to properly integrate
with the OpenStack Infrastructure. Unfortunately, there is already a
module there going by the 'arbiter' name. This switches the module name
to 'yaml2ical', which is not sounding nice, but at least is functionally
correct.

The git repository itself shall be renamed once this is merged.

Change-Id: I896bcbe7be8c6a017015b428c6ff0d39a4974a04