Fix issue with meetings that start on day1 and end on day2
Consider the following meetings (both at the same venue)
Meeting1:  Monday at 2330 for 1 hour
Meeting2: Tuesday at 0000 for 1 hour

It's clear that these meetings conflict.  The current conflicts() check
will return no conflict becuase we explictly check the start day.

Also there is a special case where day1 is Sunday so handle that case.

Fix this by converting the day into a date and including that in the conflicts
check.

Change-Id: I06f1f2ab644bf2f6aeb0e62258884a5ff85b6706